Emails: Extract Professional Email price component from Email Comparison page
