+++ client/lib/titan/is-domain-eligible-for-titan-free-trial.js
@@ -1,7 +1,11 @@
 /**
+ * @typedef { import("client/lib/domains/types").ResponseDomain } ResponseDomain domain object
+ */
+
+/**
  * Determines if the specified domain is eligible for the Titan 3-month free trial.
  *
- * @param {object} domain - domain object
+ * @param {ResponseDomain|undefined} domain - domain object
  * @returns {boolean} whether the domain is eligible or not
  */
 export function isDomainEligibleForTitanFreeTrial( domain ) {
