+++ client/my-sites/email/email-providers-comparison/price/style.scss
@@ -1,0 +1,3 @@
+professional-email-price__discount.badge {
+	margin-bottom: 12px;
+}
