+++ client/my-sites/email/email-providers-stacked-comparison/provider-cards/professional-email-card.scss
@@ -11,7 +11,3 @@
 		padding-right: 12px;
 	}
 }
-
-.professional-email-card__discount.badge {
-	margin-bottom: 12px;
-}
